+++ index.ts
@@ -1,9 +1,13 @@
+import { Rb1HandlersCommon } from "./handlers/rb1/common"
+import { Rb4HandlersCommon } from "./handlers/rb4/common"
 import { Rb5HandlersCommon } from "./handlers/rb5/common"
 import { Rb6HandlersCommon } from "./handlers/rb6/common"
 import { Rb6HandlersWebUI } from "./handlers/rb6/webui"
 import { UtilityHandlersWebUI } from "./handlers/utility/webui"
 
 export function register() {
+    R.GameCode("KBR")
+    // R.GameCode("LBR")
     R.GameCode("MBR")
 
     R.Contributor("Rox Cian", "https://github.com/RoxCian")
@@ -16,6 +20,8 @@ export function register() {
 
     routeRb6()
     routeRb5()
+    routeRb4()
+    routeRb1()
 
     R.Unhandled()
 }
@@ -42,4 +48,20 @@ function routeRb5() {
     R.Route("player.rb5_player_write_5", Rb5HandlersCommon.WritePlayer)
     R.Route("lobby.rb5_lobby_read", Rb5HandlersCommon.ReadLobby)
     R.Route("lobby.rb5_lobby_entry", Rb5HandlersCommon.ReadLobby)
+}
+
+function routeRb4() {
+    R.Route("pcb.rb4boot", Rb4HandlersCommon.BootPcb)
+    R.Route("player.rb4_player_start", Rb4HandlersCommon.StartPlayer)
+    // R.Route("player.rb5_player_read", Rb5HandlersCommon.ReadPlayer)
+    // R.Route("player.rb5_player_read_score_5", Rb5HandlersCommon.ReadPlayerScore)
+    // R.Route("player.rb5_player_read_score_old_5", Rb5HandlersCommon.ReadPlayerScoreOldVersion)
+    // R.Route("player.rb5_player_write_5", Rb5HandlersCommon.WritePlayer)
+    // R.Route("lobby.rb5_lobby_read", Rb5HandlersCommon.ReadLobby)
+    // R.Route("lobby.rb5_lobby_entry", Rb5HandlersCommon.ReadLobby)
+}
+
+function routeRb1() {
+    R.Route("pcb.rb1boot", Rb1HandlersCommon.BootPcb)
+    R.Route("player.start", Rb1HandlersCommon.StartPlayer)
 }